requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langgraph>=0.0.20
fastapi>=0.104.0
python-multipart>=0.0.6
uvicorn>=0.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pydantic>=2.4.2
python-dotenv>=1.0.0
tavily-python>=0.2.2
beautifulsoup4>=4.9.3
requests>=2.28.0
lxml>=4.9.0